Rename section titles more descriptively
Before this, the titles tried to be very close to the naming in the
diagrams.
I took some liberties as I think the reader can look at the top case of
each flow diagram to place it in context of other diagrams.
This allows to place a bit more context into the titles, which may help
the reader.