handle replay and loss: resend ExpectCandidate on timing out
Also remove multi-hop relocation with shorter_prefix_section().
Note that this can be sloted between connected and resource proof if
we need it (i.e relocate one hop at a time, but do multiple hop before
settling).

To handle replay of RPC, we ensure both source and destination agree
a node is relocated:
The destination store the candidate until it times out.
The source signs a single RelocatedInfo for a candidate, that is checked
by the target when the candidate try to join.