fix/routing_model: Simplify resource proof process to what's in 1631
We are aiming in a first phase to reduce the amount of changes we
need to do to the routing code if possible. We accept that nodes
may be lost in case of churn split while relocating as these should
happen rarely.

This also move back to a mutli hop / triangular ExpectCandidate/
ExpectCandidateResponse with its own set of issues (difficult to bound
timeout/response not from initial destination...), but that is also
closer to our code base.

Note:
1 - This mostly represent the behaviour for the related code as of
    PR 1631.
2 - This is not sufficient to guarantee loss only for churn, we still
    need to update the joining node to at least resend candidate info
    until it receives a ResourceProofResponse.
3 - There are alternative to bring back not losing nodes with churn:
    For example, add a Routing level CandidateInfo like request that
    respond with the elders to connect and the proofs that they are
    correct.
4 - There are alternative to bring back single hop communication for
    ExpectCandidate/ExpectCandidateResponse: For example, building on
    3, this Routing level CandidateInfo like request could trigger
    a new relocation, or request to wait to start resource proof.